refactor: main class 및 입출력 함수 분리
